requirements.txt
fastapi==0.116.2
uvicorn==0.35.0
asyncpg==0.30.0
sqlmodel==0.0.27
SQLAlchemy==2.0.44
pydantic==2.11.9
pydantic-settings==2.10.1
PyJWT==2.10.1
python-jose==3.5.0
bcrypt==4.0.1
passlib==1.7.4
python-dotenv==1.1.1
pinecone==7.3.0
sentence-transformers==5.1.0
PyPDF2==3.0.1
python-multipart==0.0.20
python-docx==1.2.0
pandas==2.3.2
openpyxl==3.1.5
google-generativeai==0.8.5
aiofiles==25.1.0
aiosmtplib==5.0.0
pydantic[email]

stripe==12.5.1